requirements.txt
pandas
numpy
openpyxl
odfpy
xlrd
flask
plotly
kqlmagic
pyarrow
scipy>
matplotlib
python-dotenv
